LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-004-expected.txt

line1
line2
line1
line2
line1
line2
line1
line2
line1
line2
line1
line2

FAIL #target > div 1 assert_equals:
<div data-offset-x="15">line1<br>line2</div>
offsetLeft expected 15 but got 120
PASS #target > div 2
PASS #target > div 3
PASS #target > div 4
FAIL #target > div 5 assert_equals:
<div data-offset-x="140">line1<br>line2</div>
offsetLeft expected 140 but got 35
FAIL #target > div 6 assert_equals:
<div data-offset-x="147">line1<br>line2</div>
offsetLeft expected 147 but got 42
